Remove identity and user webhook tests
